Renamed "ndt" to "legacy"
ndt7 is where the action is happening. legacy is where we write code to
not break existing things.
