+++ legacy/s2c/s2c.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/gorilla/websocket"
-	"github.com/m-lab/ndt-cloud/ndt/metrics"
-	"github.com/m-lab/ndt-cloud/ndt/protocol"
-	"github.com/m-lab/ndt-cloud/ndt/testresponder"
+	"github.com/m-lab/ndt-cloud/legacy/metrics"
+	"github.com/m-lab/ndt-cloud/legacy/protocol"
+	"github.com/m-lab/ndt-cloud/legacy/testresponder"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
