+++ legacy/testrunner.go
@@ -1,4 +1,4 @@
-package ndt
+package legacy
 
 import (
 	"fmt"
@@ -8,11 +8,11 @@ import (
 	"strings"
 
 	"github.com/gorilla/websocket"
-	"github.com/m-lab/ndt-cloud/ndt/c2s"
-	"github.com/m-lab/ndt-cloud/ndt/metrics"
-	"github.com/m-lab/ndt-cloud/ndt/protocol"
-	"github.com/m-lab/ndt-cloud/ndt/s2c"
-	"github.com/m-lab/ndt-cloud/ndt/testresponder"
+	"github.com/m-lab/ndt-cloud/legacy/c2s"
+	"github.com/m-lab/ndt-cloud/legacy/metrics"
+	"github.com/m-lab/ndt-cloud/legacy/protocol"
+	"github.com/m-lab/ndt-cloud/legacy/s2c"
+	"github.com/m-lab/ndt-cloud/legacy/testresponder"
 )
 
 const (
